refactor: Move S3 utils to backend package and remove empty utils directory
- Move core/utils/s3.py to core/backends/s3_utils.py
- Co-locate S3 utilities with the backends that use them
- Remove empty core/utils/ directory
- Update imports in DuckDB backend

Better organization by placing utilities near their usage.
All tests passing after this change.
