--- core/utils/__init__.py
@@ -1,8 +1,0 @@
-"""Utilities for the core Open FinOps system."""
-
-from .s3 import S3Utils
-
-# Import table utilities from the table_utils module
-from ..table_utils import sanitize_table_name, create_table_name
-
-__all__ = ['S3Utils', 'sanitize_table_name', 'create_table_name']
